all: Refactor authn/authz examples, add notices to subpackages
Refactored auth examples in the root package and modified
godoc for subpackages to add an Authorization section referring
to the root package. I am not sure this is the entire list of
subpackages that needs to be changed (some of the package godocs
were in auto-generated files).

Opening mostly to discuss if we can add pointer for auth{n,z}
in the subpackages.

Change-Id: Id297f79e8a83dd51f1d35bd5000da9f332049562